+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 8.0.0-alpha.2 (2025-07-14)
+
 - Adjusted cost of PLT mint/burn from 100 to 50
 - Adds `pause` function to `concordium_base::protocol_level_tokens::operations`, to support pausing/unpausing
   execution of token operations.